services/ai-service/requirements.txt
# Recommendation Service Requirements
# Python 3.9+

# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# ML & Deep Learning
torch==2.1.1
transformers==4.35.2
torchvision==0.16.1

# Vector Search
faiss-cpu==1.7.4  # Use faiss-gpu if CUDA available

# Image Processing
pillow==10.1.0
numpy==1.26.2

# Optional ML utilities
scikit-learn==1.3.2  # For color clustering in get_dominant_colors

# HTTP & Async
aiohttp==3.9.1
requests==2.31.0

# Logging & Monitoring
python-json-logger==2.0.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Production Server
gunicorn==21.2.0

# Environment & Config
python-dotenv==1.0.0

# Type hints (optional but recommended)
types-requests==2.31.0.10
